+++ alicloud/extension_log.go
@@ -1,0 +1,10 @@
+package alicloud
+
+type IndexFiledType string
+
+const (
+	TextType   = IndexFiledType("text")
+	LongType   = IndexFiledType("long")
+	DoubleType = IndexFiledType("double")
+	JsonType   = IndexFiledType("json")
+)